requirements.txt
streamlit>=1.28.0
scikit-learn>=1.3.0
pillow>=9.0.0
numpy>=1.21.0
streamlit-drawable-canvas>=0.9.0